Allow "url" to be used in addition to "continueUrl"
